requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.5
django-extensions==3.2.3
pillow==11.1.0
pip==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
qrcode==8.0
reportlab==4.2.5
sqlparse==0.5.3
